% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendan.R
\name{cat0}
\alias{cat0}
\alias{cat0n}
\alias{catn}
\title{Shortcut to cat with/without spaces and with/without newline}
\usage{
cat0(..., sep = "")

cat0n(..., sep = "")

catn(...)
}
\arguments{
\item{...}{parameters to be passed to \code{\link[base]{cat}}}

\item{sep}{as \code{\link[base]{cat}}}
}
\value{
None (invisible \code{NULL})
}
\description{
Shortcut to cat with/without spaces and with/without newline
}
\examples{
cat("This", "function", "makes", "spaces, ")
cat0("but", "this", "function", "does", "not.")
catn("This", "function", "ends", "with", "a", "newline.")
cat0n("So", "does", "this", "one.")
}
