% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_cols.R
\name{cols_to_matrix}
\alias{cols_to_matrix}
\alias{matrix_to_cols}
\title{Convert between a matrix column and several vector columns}
\usage{
cols_to_matrix(.data, name)

matrix_to_cols(.data, name)
}
\arguments{
\item{.data}{a \code{data.frame}.}

\item{name}{a \code{character} of length 1 giving the name of the matrix
column.  The corresponding vector columns' names are formatted as
"name.col", where "col" is the column name within the matrix.}
}
\value{
A modified version of the input \code{data.frame} with matching
  vector columns combined into a matrix column (for \code{cols_to_matrix}) or
  the matrix column separated into vector columns (for
  \code{matrix_to_cols}).
}
\description{
Convert between a matrix column and several vector columns
}
\details{
If \code{.data} is a \code{\link[tibble]{tibble}}, then
  \code{cols_to_matrix} will return a \code{\link[tibble]{tibble}} which does
  not conform to the standards, i.e., it will not pass
  \code{\link[tibble]{as_tibble}(validate = TRUE)}.  However, most
  \code{dplyr} verbs seem to work as intended. One set of exceptions is that
  it does not work to execute a join on matrix column(s). Joining on other
  columns is fine.
}
\examples{
d <- data.frame(a = sin(1:10),
                x.a = 1:10,
                x.b = 10:1,
                y = 2*(1:10)-3)
d <- cols_to_matrix(d, "x")
str(d)
d <- matrix_to_cols(d, "x")
str(d)
}
