% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brendan.R
\name{time_process}
\alias{time_process}
\title{Time a command}
\usage{
time_process(text, body, to = stdout())
}
\arguments{
\item{text}{A character string to be printed before \code{body} is executed.}

\item{body}{An expression.}

\item{to}{A connection to print the \code{text} and timing information to.

No space is printed between the message text and the time.}
}
\value{
The result of \code{expression}, invisibly.
}
\description{
Prints a status string, evaluates an expression, and then prints the elapsed time
required.
}
\examples{
time_process("Sleeping for 3 seconds...", Sys.sleep(3))
}
